+++ test/_require.js
@@ -1,0 +1,6 @@
+import fetch, {Headers} from 'node-fetch';
+
+global.window = {
+	fetch,
+	Headers
+};
